+++ release_bot/__init__.py
@@ -1,4 +1,3 @@
-from release_bot import release_bot
 from .version import __version__ as version
 
 __version__ = version
