+++ tests/test_specfile.py
@@ -1,9 +1,8 @@
-import release_bot.release_bot as release_bot
-from release_bot.release_bot import configuration
+from release_bot.configuration import configuration
 import pytest
 from pathlib import Path
 # import datetime from release_bot, because it needs to be patched
-from release_bot.release_bot import datetime
+from release_bot.utils import datetime, update_spec
 
 FAKE_TIME = datetime.datetime(2018, 12, 24, 17, 35, 55)
 
@@ -82,13 +81,13 @@ class TestSpecFile:
 
     def test_missing_spec(self, valid_new_release):
         with pytest.raises(SystemExit) as error:
-            release_bot.update_spec("", valid_new_release)
+            update_spec("", valid_new_release)
         assert error.type == SystemExit
         assert error.value.code == 1
 
     # test with no defined changelog
     def test_valid_conf(self, valid_spec, valid_new_release, spec_updated, patch_datetime_now):
-        release_bot.update_spec(valid_spec, valid_new_release)
+        update_spec(valid_spec, valid_new_release)
         with open(valid_spec) as spec, open(spec_updated) as original:
             assert spec.read() == original.read()
 
@@ -96,6 +95,6 @@ class TestSpecFile:
     def test_valid_conf_changelog(self, valid_spec, valid_new_release,
                                   spec_updated_changelog, patch_datetime_now):
         valid_new_release['changelog'] = ['Changelog entry 1', 'Changelog entry 2']
-        release_bot.update_spec(valid_spec, valid_new_release)
+        update_spec(valid_spec, valid_new_release)
         with open(valid_spec) as spec, open(spec_updated_changelog) as original:
             assert spec.read() == original.read()
